// touristagency-front/generated/models/CarColor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tourist agency
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const CarColor = {
    White: 'white',
    Black: 'black',
    Brown: 'brown',
    Yellow: 'yellow',
    Red: 'red',
    Blue: 'blue',
    Silver: 'silver'
} as const;
export type CarColor = typeof CarColor[keyof typeof CarColor];


export function CarColorFromJSON(json: any): CarColor {
    return CarColorFromJSONTyped(json, false);
}

export function CarColorFromJSONTyped(json: any, ignoreDiscriminator: boolean): CarColor {
    return json as CarColor;
}

export function CarColorToJSON(value?: CarColor | null): any {
    return value as any;
}